fixed a few bugs here and there, namely conversion i assumed it wouldn't be a problem to use longer type to make conversion but forgot about C crappy implicit promotion of integer fricking language i swear to god, anyway now I force conversion by implicitly castin and I think in the future I will try to create a macro system or functions to make conversion explicit because I'm tired of this implicit conversion bullshit

+++ ft_printf/ft_print_fmt_i8.c
@@ -12,11 +12,10 @@
 
 #include "ft_printf.h"
 
-int	
-ft_print_fmt_i8(va_list *arg)
+int	ft_print_fmt_i8(va_list *arg)
 {
 	int	i8;
 
-	i8 = va_arg(*arg, int);
+	i8 = (int)va_arg(*arg, int);
 	return (write(1, &i8, 1));
 }
